changed id type to int
git-svn-id: https://src.springframework.org/svn/spring-maintenance/trunk@3047 fd5a2b45-1f63-4059-99e9-3c7cb7fd75c8

+++ samples/petclinic/src/org/springframework/samples/petclinic/jdbc/JdbcPet.java
@@ -7,26 +7,24 @@ import org.springframework.samples.petclinic.Pet;
  */
 public class JdbcPet extends Pet {
 
-	private long typeId;
+	private int typeId;
 
-	private long ownerId;
+	private int ownerId;
 
-	public void setTypeId(long typeId) {
+	public void setTypeId(int typeId) {
 		this.typeId = typeId;
 	}
 
-	public long getTypeId() {
+	public int getTypeId() {
 		return this.typeId;
 	}
 
-	public void setOwnerId(long ownerId) {
+	public void setOwnerId(int ownerId) {
 		this.ownerId = ownerId;
 	}
 
-	public long getOwnerId() {
+	public int getOwnerId() {
 		return ownerId;
 	}
-
-
 
 }
